# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Vuelos",
        "description": "Documentación de la API para gestionar vuelos.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/itineraries": {
            "get": {
                "tags": [
                    "Itineraries"
                ],
                "summary": "Obtener todos los itinerarios",
                "operationId": "5d90235e7073a0dd32b06187499076eb",
                "responses": {
                    "200": {
                        "description": "Lista de itinerarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Itinerary"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error fetching itineraries"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detalles del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Itineraries"
                ],
                "summary": "Crear un itinerario",
                "operationId": "d7630299125a9d026d981991fe9985ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "departureCity": {
                                        "type": "string",
                                        "example": "New York"
                                    },
                                    "arrivalCity": {
                                        "type": "string",
                                        "example": "London"
                                    },
                                    "dateDeparture": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-11-15"
                                    },
                                    "timeDeparture": {
                                        "type": "string",
                                        "example": "08:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Itinerario creado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al guardar el itinerario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error saving itinerary"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detalles del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/itineraries/{id}": {
            "get": {
                "tags": [
                    "Itineraries"
                ],
                "summary": "Obtener un itinerario por ID",
                "operationId": "543512c6aaff0208b840f1c9f2da2d24",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del itinerario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del itinerario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Itinerary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Itinerario no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Itinerary not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener el itinerario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error fetching itinerary"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detalles del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Itineraries"
                ],
                "summary": "Actualizar un itinerario",
                "operationId": "3d2ee20340d4ad0450df3c3eb7b7b7b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del itinerario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "departureCity": {
                                        "type": "string",
                                        "example": "New York"
                                    },
                                    "arrivalCity": {
                                        "type": "string",
                                        "example": "London"
                                    },
                                    "dateDeparture": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-11-15"
                                    },
                                    "timeDeparture": {
                                        "type": "string",
                                        "example": "08:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Itinerario actualizado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Itinerary updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al actualizar el itinerario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error updating itinerary"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detalles del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Itineraries"
                ],
                "summary": "Eliminar un itinerario",
                "operationId": "b89ac2f7c7f4bfe82f850914f93240df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del itinerario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Itinerario eliminado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Itinerary deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al eliminar el itinerario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error deleting itinerary"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detalles del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Itinerary": {
                "title": "Itinerary Schema",
                "description": "Esquema de un itinerario",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "departureCity": {
                        "type": "string",
                        "example": "New York"
                    },
                    "arrivalCity": {
                        "type": "string",
                        "example": "London"
                    },
                    "dateDeparture": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-15"
                    },
                    "timeDeparture": {
                        "type": "string",
                        "example": "08:00:00"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}